+++ pkg/gomel/random_source.go
@@ -7,7 +7,7 @@ type RandomSource interface {
 	// Init initializes the random source with the given dag.
 	Init(Dag)
 	// RandomBytes returns random bytes for a given process and level.
-	RandomBytes(int, int) []byte
+	RandomBytes(uint16, int) []byte
 	// CheckCompliance checks whether the data included in the unit
 	// is compliant.
 	CheckCompliance(Unit) error
@@ -15,5 +15,5 @@ type RandomSource interface {
 	Update(Unit)
 	// DataToInclude returns data which should be included in a unit
 	// with the given creator and set of parents.
-	DataToInclude(creator int, parents []Unit, level int) ([]byte, error)
+	DataToInclude(creator uint16, parents []Unit, level int) ([]byte, error)
 }
